More release fixes
